pyproject.toml
[project]
name = "multi-llm-conversation"
version = "0.1.0"
description = "Multi-LLM conversation system with distinct speaker personalities"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "openai>=1.109.1",
    "rich>=14.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "ruff>=0.15.8",
]

[project.scripts]
multi-llm = "src.main:main"

[tool.ruff]
target-version = "py312"
exclude = ["3-way-llm-conversation.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.black]
target-version = ["py312"]
exclude = "3-way-llm-conversation.ipynb"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
explicit_package_bases = true
